Fix GCC warning, -Werror is back

+++ src/mc_control/ForceContactSensor.cpp
@@ -31,7 +31,7 @@ void ForceSensor::computeOffset()
   {
     off += v;
   }
-  off = off / lastValues.size();
+  off = off / static_cast<double>(lastValues.size());
   offset = off;
 }
